+++ app/src/modules/content/components/version-menu.vue
@@ -3,11 +3,11 @@ import api from '@/api';
 import { useCollectionPermissions } from '@/composables/use-permissions';
 import { getVersionDisplayName } from '@/utils/get-version-display-name';
 import { unexpectedError } from '@/utils/unexpected-error';
+import ComparisonModal from '@/views/private/components/comparison/comparison-modal.vue';
 import { ContentVersion, PrimaryKey } from '@directus/types';
 import slugify from '@sindresorhus/slugify';
 import { computed, ref, toRefs, unref, watch } from 'vue';
 import { useI18n } from 'vue-i18n';
-import ComparisonModal from './comparison-modal.vue';
 
 interface Props {
 	collection: string;
