+++ app/src/views/private/components/comparison/use-comparison.ts
